refactor(type): split action reaction key types
