+++ Source/Portfolio/Type/CActionKeyTypes.cpp
@@ -1,0 +1,7 @@
+#include "Type/CActionKeyTypes.h"
+
+bool FActionDataKey::IsValidMinimal() const
+{
+	return ActionType != EActionType::None
+		&& ActionType != EActionType::Max;
+}
