+++ Source/Portfolio/Type/CActionTypes.cpp
@@ -1,7 +1,1 @@
 #include "Type/CActionTypes.h"
-
-bool FActionDataKey::IsValidMinimal() const
-{
-	return ActionType != EActionType::None
-		&& ActionType != EActionType::Max;
-}
