+++ Source/Portfolio/Type/CReactionDataTypes.cpp
@@ -2,13 +2,6 @@
 
 #include "Reaction/CReaction.h"
 
-bool FReactionDataKey::IsValidMinimal() const
-{
-	return ReactionType != EReactionType::None
-		&& ReactionType != EReactionType::Max
-		&& DamageSpecKey.IsValidMinimal();
-}
-
 bool FReactionData::IsValidMinimal() const
 {
 	return ReactionDataKey.ReactionType != EReactionType::None
