+++ Source/Portfolio/Type/CReactionDataTypes.h
@@ -1,46 +1,9 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "Type/CReactionTypes.h"
-#include "Type/CCombatDamageTypes.h"
+#include "Type/CReactionKeyTypes.h"
 #include "Type/CExecutionRuleTypes.h"
 #include "CReactionDataTypes.generated.h"
-
-USTRUCT(BlueprintType)
-struct FReactionDataKey
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(EditAnywhere)
-	FDamageSpecKey DamageSpecKey = FDamageSpecKey();
-
-	UPROPERTY(EditAnywhere)
-	EReactionType ReactionType = EReactionType::None;
-
-public:
-	FReactionDataKey() = default;
-
-public:
-	bool IsValidMinimal() const;
-
-public:
-	bool operator==(const FReactionDataKey& InOther) const
-	{
-		return ReactionType == InOther.ReactionType
-			&& DamageSpecKey == InOther.DamageSpecKey;
-	}
-};
-
-FORCEINLINE uint32 GetTypeHash(const FReactionDataKey& InKey)
-{
-	uint32 H = 0;
-
-	H = HashCombine(H, GetTypeHash(InKey.DamageSpecKey));
-	H = HashCombine(H, GetTypeHash(static_cast<uint8>(InKey.ReactionType)));
-
-	return H;
-}
 
 USTRUCT(BlueprintType)
 struct FReactionData
