+++ Source/Portfolio/Type/CReactionKeyTypes.cpp
@@ -1,0 +1,8 @@
+#include "Type/CReactionKeyTypes.h"
+
+bool FReactionDataKey::IsValidMinimal() const
+{
+	return ReactionType != EReactionType::None
+		&& ReactionType != EReactionType::Max
+		&& DamageSpecKey.IsValidMinimal();
+}
